+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerLauncherCodeGenerator.kt
@@ -1,5 +1,6 @@
 package com.wutsi.codegen.kotlin.server
 
+import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.FileSpec
@@ -8,13 +9,16 @@ import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.TypeSpec
 import com.wutsi.codegen.Context
 import com.wutsi.codegen.kotlin.AbstractKotlinCodeGenerator
+import com.wutsi.codegen.kotlin.KotlinMapper
+import com.wutsi.codegen.model.Api
 import io.swagger.v3.oas.models.OpenAPI
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.cache.annotation.EnableCaching
 import org.springframework.scheduling.annotation.EnableAsync
 import org.springframework.scheduling.annotation.EnableScheduling
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.transaction.annotation.EnableTransactionManagement
-import java.io.File
 
 class ServerLauncherCodeGenerator : AbstractKotlinCodeGenerator() {
     companion object {
@@ -22,25 +26,27 @@ class ServerLauncherCodeGenerator : AbstractKotlinCodeGenerator() {
     }
 
     override fun generate(openAPI: OpenAPI, context: Context) {
-        generateClass(context)
+        val mapper = KotlinMapper(context)
+        val api = mapper.toAPI(openAPI)
+        generateClass(api, context)
     }
 
-    private fun generateClass(context: Context) {
+    private fun generateClass(api: Api, context: Context) {
         val directory = getSourceDirectory(context)
         val classname = ClassName(context.basePackage, CLASSNAME)
-        val relativePath = classname.toString().replace('.', File.separatorChar)
-        if (File(directory.absolutePath + File.separator + relativePath + ".kt").exists())
-            return
+//        val relativePath = classname.toString().replace('.', File.separatorChar)
+//        if (File(directory.absolutePath + File.separator + relativePath + ".kt").exists())
+//            return
 
         System.out.println("Generating $classname to $directory")
         FileSpec.builder(classname.packageName, classname.simpleName)
-            .addType(toTypeSpec(context))
+            .addType(toTypeSpec(api, context))
             .addFunction(toFunSpec())
             .build()
             .writeTo(getSourceDirectory(context))
     }
 
-    private fun toTypeSpec(context: Context): TypeSpec {
+    private fun toTypeSpec(api: Api, context: Context): TypeSpec {
         val spec = TypeSpec.classBuilder(ClassName(context.basePackage, CLASSNAME))
             .addAnnotation(SpringBootApplication::class)
             .addAnnotation(EnableAsync::class)
@@ -51,7 +57,14 @@ class ServerLauncherCodeGenerator : AbstractKotlinCodeGenerator() {
             spec.addAnnotation(EnableCaching::class.java)
         if (context.hasService(Context.SERVICE_MQUEUE))
             spec.addAnnotation(EnableScheduling::class.java)
-
+        if (api.isSecured()) {
+            spec.addAnnotation(EnableWebSecurity::class.java)
+            spec.addAnnotation(
+                AnnotationSpec.builder(EnableGlobalMethodSecurity::class.java)
+                    .addMember("prePostEnabled = true")
+                    .build()
+            )
+        }
         return spec.build()
     }