+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerMavenCodeGenerator.kt
@@ -10,16 +10,20 @@ import io.swagger.v3.oas.models.OpenAPI
 class ServerMavenCodeGenerator(private val mapper: KotlinMapper) : AbstractMavenCodeGenerator() {
     override fun getTemplatePath() = "/kotlin/server/pom.xml.mustache"
 
-    override fun toMustacheScope(openAPI: OpenAPI, context: Context) = mapOf(
-        "artifactId" to artifactId(context),
-        "sdkArtifactId" to SdkMavenCodeGenerator(mapper).artifactId(context),
-        "groupId" to context.basePackage,
-        "jdkVersion" to context.jdkVersion,
-        "version" to openAPI.info?.version,
-        "githubUser" to context.githubUser,
-        "githubProject" to context.githubProject,
-        "services" to toServices(context)
-    )
+    override fun toMustacheScope(openAPI: OpenAPI, context: Context): Map<String, Any?> {
+        val api = mapper.toAPI(openAPI)
+        return mapOf(
+            "artifactId" to artifactId(context),
+            "sdkArtifactId" to SdkMavenCodeGenerator(mapper).artifactId(context),
+            "groupId" to context.basePackage,
+            "jdkVersion" to context.jdkVersion,
+            "version" to openAPI.info?.version,
+            "githubUser" to context.githubUser,
+            "githubProject" to context.githubProject,
+            "services" to toServices(context),
+            "security" to api.isSecured()
+        )
+    }
 
     private fun toServices(context: Context): Map<String, Any?> {
         val result = mutableMapOf<String, Any?>()