+++ src/main/kotlin/com/wutsi/codegen/model/Api.kt
@@ -4,5 +4,9 @@ data class Api(
     val packageName: String,
     val name: String,
     val endpoints: List<Endpoint>,
-    val servers: List<Server> = emptyList()
-)
+    val servers: List<Server> = emptyList(),
+    val securities: List<Security> = emptyList()
+) {
+    fun isSecured(): Boolean =
+        endpoints.find { it.isSecured() } != null
+}